services/ai-agent/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
anthropic==0.40.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
geoalchemy2==0.15.2
redis[hiredis]==5.1.1
structlog==24.4.0
apscheduler==3.10.4
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
